+++ SkyWay.framework/Versions/A/Headers/SKWMediaStream.h
@@ -74,6 +74,12 @@
  */
 - (SKWCameraPositionEnum)getCameraPosition;
 
+/**
+ Switch camera.
+ @return Method result.
+ */
+- (BOOL)switchCamera;
+
 @end